cfg-graph: rename from cfg-walker
